fix(embed): apply embed theme without overriding the viewer's theme
The embed view applied its ?theme= URL param by calling the global
toggleColorScheme, which persists to the shared `color-scheme` localStorage
key. That polluted the viewer's own theme and, because useLocalStorage
broadcasts across tabs, caused a manual theme switch in the main app to be
reverted by any embed open in a background tab.

Instead, force the color scheme from the URL param at the app root for
/embed routes (forceColorScheme on both Mantine providers) without
persisting it, and drop the toggleColorScheme effect from EmbeddedApp. The
embed renders in its configured theme while the viewer's global, cross-tab
theme preference is left untouched. forceColorScheme is undefined on every
non-embed route, so the rest of the app is unchanged.

Verified in-browser: a dark embed no longer writes localStorage; switching
the main app to light while a dark embed is open in the background now
sticks; embed renders correctly for theme=light and theme=dark.
