+++ packages/frontend/src/App.tsx
@@ -8,6 +8,7 @@ import {
     CommercialWebAppRoutes,
 } from './ee/CommercialRoutes';
 import { AiAgentsGlobalProvider } from './ee/features/aiCopilot/components/Launcher/AiAgentsGlobalProvider';
+import { parseEmbedThemeParams } from './ee/providers/Embed/parseEmbedThemeParams';
 import ErrorBoundary from './features/errorBoundary/ErrorBoundary';
 import { SourceCodeEditorProvider } from './features/sourceCodeEditor';
 import ChartColorMappingContextProvider from './hooks/useChartColorConfig/ChartColorMappingContextProvider';
@@ -32,6 +33,14 @@ import Routes from './Routes';
 const isMobile = window.innerWidth < 768;
 
 const isMinimalPage = window.location.pathname.startsWith('/minimal');
+
+// On embed routes, force the color scheme from the ?theme= URL param without
+// persisting it to localStorage. This keeps the embed in its configured theme
+// while never overriding the viewer's own (shared, cross-tab) theme preference.
+// `undefined` everywhere else, so non-embed routes are unaffected.
+const embedForcedColorScheme = window.location.pathname.startsWith('/embed')
+    ? parseEmbedThemeParams().theme
+    : undefined;
 
 // Sentry wrapper for createBrowserRouter
 const sentryCreateBrowserRouter =
@@ -78,8 +87,13 @@ const App = () => (
         <DocumentTitle />
 
         <ReactQueryProvider>
-            <MantineProvider withGlobalStyles withNormalizeCSS withCSSVariables>
-                <Mantine8Provider>
+            <MantineProvider
+                withGlobalStyles
+                withNormalizeCSS
+                withCSSVariables
+                forceColorScheme={embedForcedColorScheme}
+            >
+                <Mantine8Provider forceColorScheme={embedForcedColorScheme}>
                     <ModalsProvider>
                         <RouterProvider router={router} />
                     </ModalsProvider>
